update useEdgeGeometry to include selfLoop scenario
